feat: 实现通知列表查询接口
- 新增 NotifyMsgVO 视图对象，支持枚举类型的 API 响应
- 创建 NotifyController 提供通知列表查询端点
- 更新 NotifyMsgService 返回 VO 而不是 DTO
- 优化服务实现填充用户信息和枚举转换
- 完善 DAO 层分页查询方法

+++ bytelogs-service/src/main/java/top/harrylei/forum/service/notify/repository/dao/NotifyMsgDAO.java
@@ -24,7 +24,7 @@ public class NotifyMsgDAO extends ServiceImpl<NotifyMsgMapper, NotifyMsgDO> {
      * @param pageSize 页大小
      * @return 通知列表
      */
-    public Page<NotifyMsgDO> pageByUserId(Long userId, Long pageNum, Long pageSize) {
+    public Page<NotifyMsgDO> pageByUserId(Long userId, Integer pageNum, Integer pageSize) {
         return lambdaQuery()
                 .eq(NotifyMsgDO::getNotifyUserId, userId)
                 .orderByDesc(NotifyMsgDO::getId)
